FXIOS-1246 ⁃ Fixes #7523 - Cleanup onboarding experiment (#7685)
Go with winner of Onboarding AB test
- Version1 of onboarding card
- Version2 of sync card
