+++ Client/Application/AppDelegate.swift
@@ -221,8 +221,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UIViewControllerRestorati
 
         pushNotificationSetup()
 
-        // Leanplum user research variable setup for onboarding research
-        _ = OnboardingUserResearch()
         // Leanplum user research variable setup for New tab user research
         _ = NewTabUserResearch()
         // Leanplum setup
