+++ Client/Frontend/Intro/IntroScreenWelcomeView.swift
@@ -7,7 +7,7 @@ import UIKit
 import SnapKit
 import Shared
 
-class IntroScreenSyncViewV1: UIView, CardTheme {
+class IntroScreenWelcomeView: UIView, CardTheme {
     // Private vars
     private var fxTextThemeColour: UIColor {
         // For dark theme we want to show light colours and for light we want to show dark colours
@@ -19,21 +19,30 @@ class IntroScreenSyncViewV1: UIView, CardTheme {
     // Orientation independent screen size
     private let screenSize = DeviceInfo.screenSizeOrientationIndependent()
     // Views
-    private lazy var titleImageViewPage2: UIImageView = {
-        let imgView = UIImageView(image: UIImage(named: "tour-Sync"))
+    private lazy var titleImageViewPage1: UIImageView = {
+        let imgView = UIImageView(image: UIImage(named: "tour-Welcome"))
         imgView.contentMode = .center
         imgView.clipsToBounds = true
         return imgView
     }()
-    private lazy var subTitleLabelPage2: UILabel = {
+    private lazy var titleLabel: UILabel = {
+        let label = UILabel()
+        label.text = Strings.CardTitleWelcome
+        label.textColor = fxTextThemeColour
+        label.font = UIFont.systemFont(ofSize: 32, weight: .bold)
+        label.textAlignment = .center
+        label.adjustsFontSizeToFitWidth = true
+        return label
+    }()
+    private lazy var subTitleLabelPage1: UILabel = {
         let fontSize: CGFloat = screenSize.width <= 320 ? 16 : 20
         let label = UILabel()
-        label.text = Strings.CardTextSync
+        label.text = Strings.CardTextWelcome
         label.textColor = fxTextThemeColour
         label.font = UIFont.systemFont(ofSize: fontSize)
         label.textAlignment = .center
         label.adjustsFontSizeToFitWidth = true
-        label.numberOfLines = 3
+        label.numberOfLines = 2
         return label
     }()
     private var closeButton: UIButton = {
@@ -72,13 +81,13 @@ class IntroScreenSyncViewV1: UIView, CardTheme {
         button.titleLabel?.textAlignment = .center
         return button
     }()
-    private lazy var startBrowsingButton: UIButton = {
+    private lazy var nextButton: UIButton = {
         let button = UIButton()
-        button.setTitle(Strings.StartBrowsingButtonTitle, for: .normal)
+        button.setTitle(Strings.IntroNextButtonTitle, for: .normal)
         button.titleLabel?.font = UIFont.systemFont(ofSize: 18, weight: .semibold)
         button.setTitleColor(UIColor.Photon.Blue50, for: .normal)
         button.titleLabel?.textAlignment = .center
-        button.accessibilityIdentifier = "startBrowsingOnboardingButton"
+        button.accessibilityIdentifier = "nextOnboardingButton"
         return button
     }()
     // Helper views
@@ -91,7 +100,7 @@ class IntroScreenSyncViewV1: UIView, CardTheme {
     var signUpClosure: (() -> Void)?
     var signInClosure: (() -> Void)?
     // Basic variables
-    private var currentPage = 1
+    private var currentPage = 0
     
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
@@ -119,19 +128,24 @@ class IntroScreenSyncViewV1: UIView, CardTheme {
         }
         
         main2panel.addArrangedSubview(imageHolder)
-        imageHolder.addSubview(titleImageViewPage2)
-        titleImageViewPage2.snp.makeConstraints { make in
+        imageHolder.addSubview(titleImageViewPage1)
+        titleImageViewPage1.snp.makeConstraints { make in
             make.edges.equalToSuperview()
         }
         
         main2panel.addArrangedSubview(bottomHolder)
-        [subTitleLabelPage2, signUpButton, signInButton, startBrowsingButton].forEach {
+        [titleLabel, subTitleLabelPage1, signUpButton, signInButton, nextButton].forEach {
              bottomHolder.addSubview($0)
-         }
+        }
         
-        subTitleLabelPage2.snp.makeConstraints { make in
-            make.left.right.equalToSuperview().inset(35)
+        titleLabel.snp.makeConstraints { make in
+            make.left.right.equalToSuperview().inset(10)
             make.top.equalToSuperview()
+        }
+        
+        subTitleLabelPage1.snp.makeConstraints { make in
+            make.left.right.equalToSuperview().inset(35)
+            make.top.equalTo(titleLabel.snp.bottom)
         }
         
         let buttonEdgeInset = 15
@@ -147,18 +161,17 @@ class IntroScreenSyncViewV1: UIView, CardTheme {
         signInButton.addTarget(self, action: #selector(showEmailLoginFlow), for: .touchUpInside)
         signInButton.snp.makeConstraints { make in
             make.left.right.equalToSuperview().inset(buttonEdgeInset)
-            make.bottom.equalTo(startBrowsingButton.snp.top).offset(-buttonSpacing)
+            make.bottom.equalTo(nextButton.snp.top).offset(-buttonSpacing)
             make.height.equalTo(buttonHeight)
         }
-        startBrowsingButton.addTarget(self, action: #selector(startBrowsing), for: .touchUpInside)
-        startBrowsingButton.snp.makeConstraints { make in
+        nextButton.addTarget(self, action: #selector(nextAction), for: .touchUpInside)
+        nextButton.snp.makeConstraints { make in
             make.left.right.equalToSuperview().inset(buttonEdgeInset)
             // On large iPhone screens, bump this up from the bottom
             let offset: CGFloat = UIDevice.current.userInterfaceIdiom == .pad ? 20 : (screenSize.height > 800 ? 60 : 20)
             make.bottom.equalToSuperview().inset(offset)
             make.height.equalTo(buttonHeight)
         }
-        
         addSubview(closeButton)
         closeButton.addTarget(self, action: #selector(startBrowsing), for: .touchUpInside)
         closeButton.snp.makeConstraints { make in
@@ -170,6 +183,7 @@ class IntroScreenSyncViewV1: UIView, CardTheme {
         } else {
             closeButton.tintColor = .black
         }
+
     }
     
     // MARK: Button Actions
@@ -191,8 +205,11 @@ class IntroScreenSyncViewV1: UIView, CardTheme {
         signUpClosure?()
     }
     
+    @objc private func nextAction() {
+        nextClosure?()
+    }
+    
     @objc private func dismissAnimated() {
-        startBrowsing()
         closeClosure?()
     }
 }
