+++ Client/UserResearch/SharedUserResearch.swift
@@ -7,12 +7,6 @@ import Leanplum
 import Shared
 
 struct LPVariables {
-    // Variable Used for AA test
-    static var showOnboardingScreenAA = LPVar.define("showOnboardingScreen", with: true)
-    // Variable Used for AB test
-    static var showOnboardingScreenAB = LPVar.define("showOnboardingScreen_2", with: true)
-    // Variable Used for 2nd Iteration of Onboarding AB Test
-    static var onboardingABTestV2 = LPVar.define("onboardingABTestV2", with: true)
     // Variable Used for New Tab Button AB Test
     static var newTabButtonABTest = LPVar.define("newTabButtonABTestProd", with: false)
 }
